Adressing review comments
For pull request https://github.com/bcgov/moh-hni-esb/pull/11
